+++ examples/model_compression/PP-MiniLM/ofa/run_ofa.sh
@@ -5,6 +5,7 @@ export PRE_EPOCHS=$4
 export SEQ_LEN=$5
 export CUDA_VISIBLE_DEVICES=$6
 export STUDENT_DIR=$7
+export WIDTH_LIST=$8
 
 python -u ./run_ofa.py --model_type ernie \
           --model_name_or_path ${STUDENT_DIR} \
@@ -16,4 +17,5 @@ python -u ./run_ofa.py --model_type ernie \
           --save_steps 100     \
           --output_dir ./ofa_models/$TASK_NAME/0.75/best_model/ \
           --device gpu  \
-          --width_mult_list 0.75
+          --width_mult_list ${WIDTH_LIST}
+
