+++ examples/model_compression/PP-MiniLM/quantization/quant_post.py
@@ -18,15 +18,17 @@ import os
 import time
 import argparse
 from functools import partial
+
 import numpy as np
-
 import paddle
-from paddle.metric import Accuracy
 
-import paddlenlp
 import paddleslim
 from paddlenlp.data import Stack, Tuple, Pad, Dict
-from paddlenlp.transformers import ErnieForSequenceClassification, ErnieTokenizer
+from paddlenlp.datasets import load_dataset
+from paddlenlp.transformers import ErnieTokenizer
+
+sys.path.append("../")
+from data import convert_example, METRIC_CLASSES, MODEL_CLASSES
 
 parser = argparse.ArgumentParser()
 
@@ -37,86 +39,53 @@ parser.add_argument(
     type=str,
     default="afqmc",
     required=False,
-    help="input task model dire")
+    help="Input task model directory.")
+
+parser.add_argument(
+    "--save_model_filename",
+    type=str,
+    default="int8.pdmodel",
+    required=False,
+    help="File name of quantified model.")
+
+parser.add_argument(
+    "--save_params_filename",
+    type=str,
+    default="int8.pdiparams",
+    required=False,
+    help="File name of quantified model's parameters.")
+
+parser.add_argument(
+    "--input_model_filename",
+    type=str,
+    default="float.pdmodel",
+    required=False,
+    help="File name of float model.")
+
+parser.add_argument(
+    "--input_param_filename",
+    type=str,
+    default="float.pdiparams",
+    required=False,
+    help="File name of float model's parameters.")
+
+parser.add_argument(
+    "--tokenizer_path",
+    default='../general_distill/ernie-batchbatch-50w_400000/',
+    type=str,
+    help="The directory for tokenizer.", )
 
 args = parser.parse_args()
 
-METRIC_CLASSES = {
-    "afqmc": Accuracy,
-    "tnews": Accuracy,
-    "iflytek": Accuracy,
-    "ocnli": Accuracy,
-    "cmnli": Accuracy,
-    "cluewsc2020": Accuracy,
-    "csl": Accuracy,
-}
-
-MODEL_CLASSES = {"ernie": (ErnieForSequenceClassification, ErnieTokenizer), }
-
-
-def convert_example(example,
-                    tokenizer,
-                    label_list,
-                    max_seq_length=512,
-                    is_test=False):
-    """convert a glue example into necessary features"""
-    if not is_test:
-        # `label_list == None` is for regression task
-        label_dtype = "int64" if label_list else "float32"
-        # Get the label
-        label = example['label']
-        label = np.array([label], dtype=label_dtype)
-    # Convert raw text to feature
-    if 'sentence' in example:
-        example = tokenizer(example['sentence'], max_seq_len=max_seq_length)
-    elif 'sentence1' in example:
-        example = tokenizer(
-            example['sentence1'],
-            text_pair=example['sentence2'],
-            max_seq_len=max_seq_length)
-    elif 'keyword' in example:  # CSL
-        sentence1 = " ".join(example['keyword'])
-        example = tokenizer(
-            sentence1, text_pair=example['abst'], max_seq_len=max_seq_length)
-    elif 'target' in example:  # wsc
-        text, query, pronoun, query_idx, pronoun_idx = example['text'], example[
-            'target']['span1_text'], example['target']['span2_text'], example[
-                'target']['span1_index'], example['target']['span2_index']
-        text_list = list(text)
-        assert text[pronoun_idx:(pronoun_idx + len(pronoun)
-                                 )] == pronoun, "pronoun: {}".format(pronoun)
-        assert text[query_idx:(query_idx + len(query)
-                               )] == query, "query: {}".format(query)
-        if pronoun_idx > query_idx:
-            text_list.insert(query_idx, "_")
-            text_list.insert(query_idx + len(query) + 1, "_")
-            text_list.insert(pronoun_idx + 2, "[")
-            text_list.insert(pronoun_idx + len(pronoun) + 2 + 1, "]")
-        else:
-            text_list.insert(pronoun_idx, "[")
-            text_list.insert(pronoun_idx + len(pronoun) + 1, "]")
-            text_list.insert(query_idx + 2, "_")
-            text_list.insert(query_idx + len(query) + 2 + 1, "_")
-        text = "".join(text_list)
-        example = tokenizer(text, max_seq_len=max_seq_length)
-
-    if not is_test:
-        return example['input_ids'], example['token_type_ids'], label
-    else:
-        return example['input_ids'], example['token_type_ids']
-
 
 def quant_post(args, batch_size=8, algo='avg'):
-    paddle.enable_static()
     place = paddle.set_device("gpu")
     exe = paddle.static.Executor(place)
     args.task_name = args.task_name.lower()
 
-    train_ds = paddlenlp.datasets.load_dataset(
-        "clue", args.task_name, splits="dev")
+    train_ds = load_dataset("clue", args.task_name, splits="dev")
 
-    tokenizer = ErnieTokenizer.from_pretrained(
-        "../ernie-batchbatch-50w_400000/best_models/AFQMC/")
+    tokenizer = ErnieTokenizer.from_pretrained(args.tokenizer_path)
 
     trans_func = partial(
         convert_example,
@@ -126,7 +95,7 @@ def quant_post(args, batch_size=8, algo='avg'):
         is_test=True)
     train_ds = train_ds.map(trans_func, lazy=True)
 
-    def test():
+    def batch_generator_func():
         batch_data = [[], []]
         for data in train_ds:
             batch_data[0].append(data[0])
@@ -134,12 +103,6 @@ def quant_post(args, batch_size=8, algo='avg'):
             if len(batch_data[0]) == batch_size:
                 input_ids = Pad(axis=0, pad_val=0)(batch_data[0])
                 segment_ids = Pad(axis=0, pad_val=0)(batch_data[1])
-                ones = np.ones_like(input_ids, dtype="int64")
-                seq_length = np.cumsum(ones, axis=-1)
-
-                position_ids = seq_length - ones
-                attention_mask = np.expand_dims(
-                    (input_ids == 0).astype("float32") * -1e9, axis=[1, 2])
                 yield [input_ids, segment_ids]
                 batch_data = [[], []]
 
@@ -147,13 +110,13 @@ def quant_post(args, batch_size=8, algo='avg'):
         exe,
         args.input_dir,
         os.path.join(args.task_name + '_quant_models', algo + str(batch_size)),
-        save_model_filename='int8.pdmodel',
-        save_params_filename='int8.pdiparams',
+        save_model_filename=args.save_model_filename,
+        save_params_filename=args.save_params_filename,
         algo=algo,
         hist_percent=0.9999,
-        batch_generator=test,
-        model_filename='float.pdmodel',
-        params_filename='float.pdiparams',
+        batch_generator=batch_generator_func,
+        model_filename=args.input_model_filename,
+        params_filename=args.input_param_filename,
         quantizable_op_type=['matmul', 'matmul_v2'],
         weight_bits=8,
         weight_quantize_type='channel_wise_abs_max',
